fix: #302 - 청킹 기반 관련 노트 추천 처리 보완
